Interface project/requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pillow==11.2.1
progressbar==2.5
pyparsing==3.2.3
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.16.3
shiboken6==6.10.2
six==1.17.0
tzdata==2025.2
